OTG: update methods
